producer_daily/requirements.txt
requests==2.31.0
kafka-python==2.0.2
psycopg2-binary==2.9.7
tenacity>=8.2.0